pyproject.toml
[tool.poetry]
name = "investments_f"
version = "0.1.0"
description = "investments_f"
authors = ["Filippo"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10.7"
numpy = ">=1.21.5"
pandas = ">=1.4.0"
yfinance = ">=0.1.69"
jupyterlab = "^3.4.5"
alpaca-py = "^0.6.0"

[tool.poetry.dev-dependencies]
poetry = ">=1.1.12"
pytest = ">=6.2.5"
mypy = "<=0.971"
isort = ">=5.10.1"
black = ">=21.12b0"
pylint = ">=2.12.2"
python-lsp-server = {extras = ["all"], version = ">=1.3.3"}
python-lsp-black = ">=1.0.1"
pytest-socket = ">=0.5.0"
pylsp-mypy = ">=0.5.7"
types-requests = ">=2.26.2"
ipython = ">=7.30.1"
TestSlide = ">=2.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.mypy]
ignore_missing_imports = true

[tool.pylint.master]
jobs = 0
extension-pkg-whitelist = "pydantic"

[tool.pylint.message_control]
disable = [
    "duplicate-code",
    "invalid-name",
    "invalid-unary-operand-type",
    "too-few-public-methods",
    "too-many-public-methods",
    "too-many-arguments",
    "too-many-instance-attributes",
    "too-many-locals",
    "line-too-long",
    "logging-fstring-interpolation",
    "cache-max-size-none",
    "too-many-statements",
    "trailing-whitespace",
    "trailing-newlines",
    "fixme"
]

[tool.pylint.reports]
output-format = "colorized"

[tool.pylint.logging]
logging-format-style = "new"
